sherkat/sherkat/sherkat/sherkat/requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
chardet==5.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-excel==0.0.10
django-import-export==3.0.1
django-jalali-date==1.0.1
et-xmlfile==1.1.0
jdatetime==4.1.0
kombu==5.2.4
lml==0.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
prompt-toolkit==3.0.36
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pytz==2022.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
texttable==1.6.7
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
